Implement bare command and refactor deployer.
