+++ src/main/resources/META-INF/services/io.vertx.core.spi.Command
@@ -1,2 +1,4 @@
 io.vertx.core.cli.commands.RunCommand
 io.vertx.core.cli.commands.VersionCommand
+io.vertx.core.cli.commands.BareCommand
+io.vertx.core.cli.commands.BareAliasCommand
